requirements.txt
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
websockets>=13.0
duckduckgo-search>=7.0.0
pypdf>=5.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
Pillow>=11.0.0
python-multipart>=0.0.9
